Added link to dev.to article
I found your great post via dev.to, which links to this repository, but not the other way around, so I am proposing this minor edit. I am still processing the contents of your article and I _might_ get back to you with comments, which leads me to a question - I can comment on dev.to, but I cannot seem to comment [on your blog](https://www.giftegwuenu.com/setup-continuous-integration-with-travis-ci-in-your-nodejs-app/) - I am missing something?